+++ packages/components-js/tests/e2e/helpers/puppeteer-helper.ts
@@ -59,17 +59,26 @@ export const getActiveElementTagName = async (page: Page): Promise<string> => {
   return page.evaluate(() => document.activeElement.tagName);
 };
 
-type GetElementStyleOptions = { waitForTransition: boolean };
+type Pseudo = '::before' | '::after';
+type GetElementStyleOptions = {
+  waitForTransition?: boolean;
+  pseudo?: Pseudo;
+};
 
 export const getElementStyle = async (
   element: ElementHandle,
   property: keyof CSSStyleDeclaration,
   opts?: GetElementStyleOptions
-): Promise<string> =>
-  element.evaluate(
+): Promise<string> => {
+  return await element.evaluate(
     async (el: Element, property: keyof CSSStyleDeclaration, opts?: GetElementStyleOptions): Promise<string> => {
-      const style = getComputedStyle(el);
-      if (opts?.waitForTransition) {
+      const options: GetElementStyleOptions = {
+        waitForTransition: false,
+        pseudo: null,
+        ...opts,
+      };
+      const style = getComputedStyle(el, options.pseudo);
+      if (options.waitForTransition) {
         await new Promise((resolve) => setTimeout(resolve, parseFloat(style.transitionDuration) * 1000));
       }
       return style[property].toString();
@@ -77,6 +86,48 @@ export const getElementStyle = async (
     property,
     opts
   );
+};
+
+type GetStyleOnFocusOptions = {
+  pseudo?: Pseudo;
+};
+
+export const getOutlineStyle = async (element: ElementHandle, opts?: GetStyleOnFocusOptions): Promise<string> => {
+  const options: GetStyleOnFocusOptions = {
+    pseudo: null,
+    ...opts,
+  };
+  const { pseudo } = options;
+  return `${await getElementStyle(element, 'outline', { pseudo })} ${await getElementStyle(element, 'outlineOffset', {
+    pseudo,
+  })}`;
+};
+
+export const getBoxShadowStyle = async (element: ElementHandle, opts?: GetStyleOnFocusOptions): Promise<string> => {
+  const options: GetStyleOnFocusOptions = {
+    pseudo: null,
+    ...opts,
+  };
+  const { pseudo } = options;
+  return await getElementStyle(element, 'boxShadow', { pseudo });
+};
+
+export const getStyleOnFocus = async (
+  element: ElementHandle,
+  property: 'outline' | 'boxShadow' = 'outline',
+  opts?: GetStyleOnFocusOptions
+): Promise<string> => {
+  await element.focus();
+  return property === 'outline' ? await getOutlineStyle(element, opts) : await getBoxShadowStyle(element, opts);
+};
+
+export const setAttribute = async (element: ElementHandle, key: string, value: string): Promise<void> => {
+  await element.evaluate((el, { key, value }) => el.setAttribute(key, value), { key, value });
+};
+
+export const waitForInheritedCSSTransition = async (page: Page): Promise<void> => {
+  await page.waitForTimeout(500);
+};
 
 export const getElementIndex = async (element: ElementHandle, selector: string): Promise<number> =>
   element.evaluate(async (el: Element, selector: string): Promise<number> => {