+++ packages/partials/CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### [Unreleased]
 
+#### Added
+- Extend options of `getPorscheDesignSystemCoreStyles(opts)` by `{ prefix: string }`
+
 ### [2.2.0] - 2020-09-11
 
 #### Added